Frontend/patient-portal/backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
psycopg2-binary
python-dotenv
bcrypt
